A cute osd messager :), ripped from kpdf, much better than status bar I think.
I missed a message from my friend yesterday because of the autoReply, so I make the osd to show some words when a message is autoReplyed. these words will not disappear automatically, you have to manually click it to hide it.

And I add a m_bMouseClicked to detect if the mouseRelease event is from a click or some other windows, because the osd of some app, amarok and now qterm, will hide in the mousePress event, and qterm will get the mouseRelease event. that is not we want.

TODO:
Add some configure option to control the osd maybe.


git-svn-id: https://svn.code.sf.net/p/qterm/code/trunk/qterm@209 bd9efd21-7b0e-0410-bb59-ddbc1b2ac886