Dirty hack to hide tabbar in fullscreen mode
I do not know if there is an easy way to handle this, the tabbar cannot
be hided in the tabbed view mode, so in the fullscreen mode, switch
to the subwindow mode to hide the tabbar. During the switch, I need to
remaximize the subwindow. This somehow triggered an infinite event loop in
the code when there are more than one subwindow. I also need to set the
frameless flag of the subwindows. In the long future we probably should
subclass the QMdiArea class to hide the tabbar, this will also open the
possibility to put the download progress bar to the right corner of the
tabbar.

git-svn-id: https://svn.code.sf.net/p/qterm/code/trunk/qterm-qt4@1290 bd9efd21-7b0e-0410-bb59-ddbc1b2ac886