Need to load the ciphers for openssl 1.0
Otherwise QTerm cannot read private key protected by passphrase. I do
not know if it is the correct way to fix it, but adding an
OpenSSL_add_all_ciphers certainly does the trick. I also do not know if
there is a similar problem with openssl 1.1, according to the document
everything should be loaded automatically. The worst scenario adding
an OPENSSL_init_crypto should fix it.