Adjust the layout to fit smaller screen
I have to remove all the items in the combobox from the ui file,
otherwise the layout refuse to shrink, and always grow to certain size
(depending on the system font size I guess). Also the size policy cannot
be "Preferred" otherwise it will expand itself as well. Now the default
size is 820x710, hopefully it is good enough for most of the screens.